attempting to figure out githubactions and building automagically
