--- README.md
@@ -1,3 +1,0 @@
-# ponglove2d
-pong in love2d 11.4
-:)
