+++ build.bat
@@ -1,0 +1,1 @@
+copy /b love.exe+%1 "%~n1.exe"
