--- conf.lua
@@ -1,4 +1,0 @@
-function love.conf(t)
-    t.window.width = 1024
-    t.window.height = 800
-end
