+++ src/conf.lua
@@ -1,0 +1,4 @@
+function love.conf(t)
+    t.window.width = 1024
+    t.window.height = 800
+end
